+++ stdnet/utils/__init__.py
@@ -31,6 +31,11 @@ def iterpair(iterable):
     else:
         return iterable
     
+def int_or_float(v):
+    v = float(v)
+    i = int(v)
+    return i if v==i else v
+
 def grouper(n, iterable, padvalue=None):
     "grouper(3, 'abcdefg', 'x') --> ('a','b','c'), ('d','e','f'), ('g','x','x')"
     return zip_longest(*[iter(iterable)]*n, fillvalue=padvalue)
@@ -61,20 +66,17 @@ def _flat2d_gen(iterable):
     for v in iterable:
         yield v[0]
         yield v[1]
-        
-        
+                
 def flat2d(iterable):
     if hasattr(iterable,'__len__'):
         return chain(*iterable)
     else:
         return _flat2d_gen(iterable)
-
     
 def _flatzsetdict(kwargs):
     for k,v in iteritems(kwargs):
         yield v
         yield k
-
 
 def flatzset(iterable = None, kwargs = None):
     if iterable: